refactor(desktop): extract useChannelIntro to keep ChannelPane under size ceiling
Rebase onto main brought back the Browse channels Welcome card (#1948);
the intro builder now lives in its own hook file.
